add "init.h"
